+++ bin/demyx-init.sh
@@ -15,6 +15,11 @@ if [[ -n "$(ls /run | grep docker.sock)" || -n "$DOCKER_HOST" ]]; then
     # Start the API if DEMYX_STACK_SERVER_API has a url defined (Ex: api.domain.tld)
     DEMYX_STACK_SERVER_API="$(demyx info stack --filter=DEMYX_STACK_SERVER_API --quiet)"
     [[ "$DEMYX_STACK_SERVER_API" != false ]] && demyx-api &
+
+    # Refresh stack if .env exists
+    if [[ -f "$DEMYX_STACK"/.env ]]; then
+        demyx stack refresh &
+    fi
 fi
 
 # Run sshd