+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Use jq to get Bedrock version
 - Use curly brackets for OLS variable
 - MOTD stack install message using wrong command
+- Move stack refresh command to demyx-init.sh
 - chroot.sh
     - demyx update now shows updates if available
     - Show an update notice if available